+++ packages/dashboards/examples/src/components/prototypes/data/table/acme-table-mock-data-source.service.ts
@@ -23,6 +23,7 @@ import { Inject, Injectable } from "@angular/core";
 import { BehaviorSubject, Subject } from "rxjs";
 
 import {
+    ClientSideDataSource,
     IDataField,
     INovaFilters,
     LocalFilteringDataSource,
@@ -73,32 +74,13 @@ export class AcmeTableMockDataSource extends LocalFilteringDataSource<BasicTable
 
         return new Promise((resolve) => {
             setTimeout(async () => {
-                const virtualScrollFilter =
-                    filters.virtualScroll && filters.virtualScroll.value;
-
-                if (virtualScrollFilter) {
-                    // The multiplier used here is a way to fetch more items per scroll
-                    const start = filters.virtualScroll?.value.start;
-                    const end = filters.virtualScroll?.value.end;
-                    // Note: We should start with a clean cache every time first page is requested
-                    if (start === 0) {
-                        this.cache = [];
-                    }
-                    const nextChunk = TABLE_DATA.slice(start, end);
-                    // We identify here whether the cached array does already contain some of the fetched data.
-                    // Then we update the cached array with the only values it doesn't contain
-                    this.cache = this.cache.concat(
-                        nextChunk.filter((item) => !this.cache.includes(item))
-                    );
-                    super.setData(this.cache);
-                }
+                // Set the data to the table
+                super.setData(TABLE_DATA);
 
                 const filteredData = await super.getFilteredData(filters);
-
                 if (filteredData.paginator) {
                     filteredData.paginator.total = TABLE_DATA.length;
                 }
-
                 resolve({
                     ...filteredData,
                     dataFields: this.dataFields,