+++ packages/dashboards/src/lib/components/table-widget/types.ts
@@ -46,7 +46,11 @@ export interface ITableWidgetConfig {
     columns: Array<ITableWidgetColumnConfig>;
     sortable?: boolean;
     sorterConfiguration: ITableWidgetSorterConfig;
-    hasVirtualScroll: boolean;
+    /**
+     * @deprecated Use scrollType and set it to "infinite" instead
+     */
+    hasVirtualScroll?: boolean;
+    scrollType?: ScrollType;
     interactive?: boolean;
     headerTooltipsEnabled?: boolean;
     scrollActivationDelayMs?: number;
@@ -62,9 +66,29 @@ export interface ITableWidgetConfig {
         searchTerm?: string;
         searchDebounce?: number;
     };
+
+    paginatorConfiguration?: ITableWidgetPaginatorConfig;
 }
 
 export interface ITableWidgetSorterConfig {
     descendantSorting: boolean;
     sortBy: string;
+}
+
+export interface ITableWidgetPaginatorConfig {
+    pageSizeSet?: number[];
+    pageSize?: number;
+}
+
+export interface IPaginatorState {
+    page: number;
+    pageSize: number;
+    pageSizeSet: number[];
+    total: number;
+}
+
+export enum ScrollType {
+    virtual = "virtual",
+    paginator = "paginator",
+    default = "default",
 }