+++ packages/dashboards/src/lib/configurator/components/widgets/table/filters-editor/table-filters-editor.component.ts
@@ -99,6 +99,7 @@ export class TableFiltersEditorComponent
         const descendantSortingFormControl = this.form
             .get("sorterConfiguration")
             ?.get("descendantSorting");
+
         if (changes.sorterConfiguration) {
             const sortedColumn = this.sortableColumns.find(
                 (column) => column.id === this.sorterConfiguration?.sortBy
@@ -132,6 +133,7 @@ export class TableFiltersEditorComponent
                 descendantSortingFormControl?.enable();
             }
         }
+
         this.changeDetector.detectChanges();
     }