+++ packages/dashboards/src/lib/configurator/configurator.module.ts
@@ -133,6 +133,8 @@ import { KpiWidgetColorService } from "./services/kpi-widget-color.service";
 import { ConfiguratorHeadingService } from "./services/public-api";
 import { WidgetClonerService } from "./services/widget-cloner.service";
 import { WidgetEditorService } from "./services/widget-editor.service";
+import { TableScrollTypeEditorComponent } from "./components/widgets/table/scrollType-editor/scroll-type-editor.component";
+import { ScrollTypeEditorService } from "./components/widgets/table/scrollType-editor/scroll-type-editor.service";
 /* eslint-enable max-len */
 
 const entryComponents: IComponentWithLateLoadKey[] = [
@@ -149,6 +151,7 @@ const entryComponents: IComponentWithLateLoadKey[] = [
     ProportionalChartOptionsEditorComponent,
     ProportionalChartOptionsEditorV2Component,
     TableFiltersEditorComponent,
+    TableScrollTypeEditorComponent,
     TimeseriesMetadataConfigurationComponent,
     TimeseriesSeriesCollectionConfigurationComponent,
     TableColumnsConfigurationComponent,
@@ -252,6 +255,7 @@ const exportedDeclarations = [
         KpiWidgetColorService,
         TimeseriesChartPresetService,
         TimeseriesScalesService,
+        ScrollTypeEditorService,
     ],
     exports: exportedDeclarations,
 })